Revert "Merge branch 'master' of https://github.com/Bcespedes04/WebPatrones"
This reverts commit 6bbd467cfc9f89c709e35b01e1650c9b8318f337, reversing
changes made to 6f8045caf128984680b3896ad7593cd440a5c625.

+++ src/main/java/com/proyecto/service/EventoService.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface EventoService {
     List<Evento> getEventos();
-    Evento getEvento(int id);
+    Evento getEvento(Long id);
     void save(Evento evento);
-    void delete(int id);
+    void delete(Long id);
 }
